Implemented functionality for Doctor users to see the list of available medicine and add new entries

+++ Patient Management Model/src/PatientManagement/Model/Medicines/CapsuleMedicine.java
@@ -12,7 +12,7 @@ package PatientManagement.Model.Medicines;
 public class CapsuleMedicine extends Medicine
 {
     
-    public CapsuleMedicine(String name, String description, int amount, double price)
+    public CapsuleMedicine(String name, String description, int amount, float price)
     {
         super(name, description, amount, price);
     }
